requirements.txt
flask
lxml
qrcode
pillow
pdfplumber
